#55 added limit to coordinates

+++ mechanics-module/src/main/java/marowak/dev/dto/motion/PlayerMotion.java
@@ -2,8 +2,8 @@ package marowak.dev.dto.motion;
 
 public record PlayerMotion(
         String playerName,
-        int x,
-        int y,
+        long x,
+        long y,
         int angle,
         int speed
 ) {
